pyproject.toml
[project]
name = "project-hse-phyton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "asyncio>=4.0.0",
    "python-dotenv>=1.1.1",
]
